datapath: Allow masks for set actions.
Masked set action allows more megaflow wildcarding.  Masked set action
is now supported for all writeable key types, except for the tunnel
key.

It is not clear whether masked set is useful for skb_priority.
However, we already use the LSB of pkt_mark for IPSec in tunnels, so
it might be useful to be able to set individual bits on pkt_mark.

Signed-off-by: Jarno Rajahalme <jrajahalme@nicira.com>