testsuite: Drain the list of jobs in the shell
SUS says:
    When jobs reports the termination status of a job,
    the shell removes its process ID from the list of those
    "known in the current shell execution environment";

With NetBSD /bin/sh, the list involves zombie processes and
eventually ends up with "can not fork" errors.

I guess the problem itself is not new.  Probably it just means the
growth of the number of our test cases reached the point where
it actually causes the problem in my environment.

Signed-off-by: YAMAMOTO Takashi <yamamoto@ovn.org>